+++ src/cpp/src/clip.hpp
