+++ src/cpp/src/vlm_minicpmv.hpp
@@ -1,0 +1,8 @@
+#ifndef MINICPMV_H
+#define MINICPMV_H
+
+#include "clip.hpp"
+
+std::pair<int, int> find_best_resize(std::pair<int, int> original_size, int scale_resolution, int patch_size, bool allow_upscale=false);
+std::pair<int, int> get_refine_size(std::pair<int, int> original_size, std::pair<int, int> grid, int scale_resolution, int patch_size, bool allow_upscale);
+#endif
