+++ src/cpp/src/vlm_sampling.hpp
